+++ app/Http/Controllers/Admin/UsersController.php
@@ -18,6 +18,9 @@ class UsersController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('user_access')) {
+            return abort(401);
+        }
 
 
                 $users = User::all();
@@ -32,6 +35,9 @@ class UsersController extends Controller
      */
     public function create()
     {
+        if (! Gate::allows('user_create')) {
+            return abort(401);
+        }
         
         $roles = \App\Role::get()->pluck('title', 'id');
 
@@ -47,6 +53,9 @@ class UsersController extends Controller
      */
     public function store(StoreUsersRequest $request)
     {
+        if (! Gate::allows('user_create')) {
+            return abort(401);
+        }
         $user = User::create($request->all());
         $user->role()->sync(array_filter((array)$request->input('role')));
 
@@ -67,6 +76,9 @@ class UsersController extends Controller
      */
     public function edit($id)
     {
+        if (! Gate::allows('user_edit')) {
+            return abort(401);
+        }
         
         $roles = \App\Role::get()->pluck('title', 'id');
 
@@ -85,6 +97,9 @@ class UsersController extends Controller
      */
     public function update(UpdateUsersRequest $request, $id)
     {
+        if (! Gate::allows('user_edit')) {
+            return abort(401);
+        }
         $user = User::findOrFail($id);
         $user->update($request->all());
         $user->role()->sync(array_filter((array)$request->input('role')));
@@ -120,16 +135,19 @@ class UsersController extends Controller
      */
     public function show($id)
     {
+        if (! Gate::allows('user_view')) {
+            return abort(401);
+        }
         
         $roles = \App\Role::get()->pluck('title', 'id');
-$contacts = \App\Contact::where('user_id', $id)->get();$clinics = \App\Clinic::whereHas('users',
+$contacts = \App\Contact::where('user_id', $id)->get();$api_tests = \App\ApiTest::where('created_by_id', $id)->get();$clinics = \App\Clinic::whereHas('users',
                     function ($query) use ($id) {
                         $query->where('id', $id);
                     })->get();$tasks = \App\Task::where('user_id', $id)->get();
 
         $user = User::findOrFail($id);
 
-        return view('admin.users.show', compact('user', 'contacts', 'clinics', 'tasks'));
+        return view('admin.users.show', compact('user', 'contacts', 'api_tests', 'clinics', 'tasks'));
     }
 
 
@@ -141,6 +159,9 @@ $contacts = \App\Contact::where('user_id', $id)->get();$clinics = \App\Clinic::w
      */
     public function destroy($id)
     {
+        if (! Gate::allows('user_delete')) {
+            return abort(401);
+        }
         $user = User::findOrFail($id);
         $user->delete();
 
@@ -154,6 +175,9 @@ $contacts = \App\Contact::where('user_id', $id)->get();$clinics = \App\Clinic::w
      */
     public function massDestroy(Request $request)
     {
+        if (! Gate::allows('user_delete')) {
+            return abort(401);
+        }
         if ($request->input('ids')) {
             $entries = User::whereIn('id', $request->input('ids'))->get();
